Added peripheral invalidation upon server reload
This allows to update peripheral code without having to restart the server &/or the entire game when playing locally. This seems to be the only "clean" way to do it, since CC doesn't propose a feature to remove/unregister peripherals. Although one could argue it could be done by emptying PeripheralProviderBase.peripherals (and it could work), I'm open to discuss it further.

In order to reload properly, the server owner must execute two commands in that order:
/kubejs reload startup_scripts
then
/reload
then restart the CC computer/turtle